Auto-collect script tags by giving pattern for their ids.
Now one can pass an arument called "scripts" to createCindy.  This is a
string which must contain an asterisk ('*').  That asterisk is renamed by
the event (without the "script" suffix) and if a script tag with that id
does exist, it is used as the source code for that event.

Example: passing “'scripts':'foo*bar'” to the call, a tag with
id="foomousedragbar" will be used for mouse drag events, if it exists.