Automatically wait for GWT modules loaded before CindyJS.
We look at the list of active modules, and for each active module wait for a
message that said module has been loaded successfully.  The createCindy call
is deferred until the final waited-for object has arrived.  As a by-product,
we wait for a complete DOM load on other (i.e. non-GWT) pages as well.