Implement scaleAndOrigin transformation.
This transformation sets the scale and the origin position, disregarding the
initial transformation.  It can be used to easily apply transformations read
from CDY files, without relying too much on the exact makeup of the initial
transformation.  Note that the Y position of the origin is in pixels from
the bottom up, not from the top down as it is in Cinderella.