Handle undeclared variables reported by JSHint.
This takes care of all the W117 undeclared variables (including assignment
to such, which would be assignments to the global window object).  These
should be the most likely sources of trouble with "use strict".  Some few
cases remain, in imported code.  But these cases have been checked manually
and are acceptable.