Drive-by fix: always return nada instead of unset.
I noticed that a lot of code in Operators does “=== nada” comparisons.
Having two instances of the Nada type breaks those.  So we make sure that
the “unset” object is only used inside variables, and never returned.  The
unset object might at this point be any object, having it be an instance of
Nada is just a semantically sensible thing, but no real requirement.