Disable susplicious function numeric.T.eig.
This function looks in part like a class (i.e. static method), and in part
like an instance method, but neither makes too much sense as it stands.
Reported upstream as https://github.com/sloisel/numeric/issues/61
Will disable the method for now, since it is unusable in its current form.
Disabling it gets rid of a bunch of closure compiler warnings.