Start implementing Cindy3D for CindyJS.
The Shader code is copied from the Cindy3D project,
but adjusted for WebGL with frag_depth extension.

The port is compiled using advanced compilation of the closure compiler.
The resulting file is loaded separately, and will register its functions
with the core evaluator of CindyJS. The example demonstrates usage.