Include arity in function name. Allow redefining functions.
At compile time, the number of arguments is included into the name of the
function to call.  This avoids argument count checking all over the place.
User-defined functions can again overwrite built-in ones of the same arity.
Legacy code is in place to allow an incremental migration to the new
convention.  At the moment, no operators are converted to the new style.