Create a Plugin system.
Plugins are either registered globally using createCindy.registerPlugin or
locally using the plugins parameter to a createCindy call.  In either case,
a script may use a plugin by invoking the "use" function.  In that case, the
plugin initialization callback will be invoked and the plugin can define
additional functions for the instance which is being created.