Restructure startup and introduce shutdown procedures.
Now the created instances have a shutdown function to terminate them.
A DOM Mutation Observer will help detecting removal of the canvas
and shut the instance down automatically in that event.
Upon shutdown, all event listeners are removed and the timer stops.

The code waiting for loading of certain DOM events during startup has been
moved outside the main closure.  To make code in that position more
consistent with the rest of the code, we now do a two-stage closure
compilation, once for the outside wrapper and once for the rest.