Turn pages files and whiteboard content into TaskPaper text.
This was done in part automatically using the added Perl script, and in part
manually or with the help of Emacs. Colors were mapped as follows:
green -> @done, red -> @tricky, blue -> @rethink.
The existing Pages files were renamed to avoid accidential divergence.
In the long run, the Pages files should be removed and the number of
Taskpaper files might be reduced by combining them somehow.