Get rid of calls to clone.
CindyScript data types (those objects with a cstype field) are now
considered immutable.  It is an error to modify any such object after it has
been created.  For this reason, there is no need to create clones of such
objects: If the object cannot change, there is no difference between the
original and the clone.