Make canvas focusable for keydown event.
If there is no document-wide key listener and there is a keydown script,
then we make the canvas focusable and install a key listener for it.
This should do the right thing by default for most applications, although
it might be surprising in situations where users might expect some response
to key presses even if they did not focus the canvas first.