New operators: combinations and mincostmatching
The mincostmatching operator was originally intended to match objects while
tracing, but will probably perform worse than a naive iteration over all
possible permutations, since we'll be tracing at most 4 solutions and 4!
isn't that much.
But since I had the code in hand, and the Yahtzee example as well,
I decided to keep going and make that operator available nonetheless.

The combinations operator was required to make the bonus computation
work in that example. The side effect of computing a binomial coefficient
if the first argument is a number is not used in the example,
but tested by a unit test from the reference manual.