Replace “eval” with “new Function”.
The eval function returns its code in the scope which contains the eval.
This means that the code would suddenly have access to all the minified
CindyJS internals which we have otherwise carefully hidden away in out
closure.  It might prevent some compilers from optimizing all that internal
stuff more aggressively.

A new function created using the function constructor, on the other hand,
behaves as if it were contained in the global scope.  It has no special
access privileges.  To increase usefulness and avoid dirty hacks, we pass
the public interface to the current CindyJS instance as “this” argument.