Make point on segment movable again
The parameter of that operation has been changed to something which should
be reasonably stable when tracing.  As long as no endpoint of the segment
becomes a point at infinity, in which case the segment is ill defined in any
case.

I also removed el.sx, el.sy and el.sz in many places, since always
maintaining real and dehomogenized coordinates in addition to the
homogeneous ones was too much effort in the long run, and bound to fail if
things become complex.

csinit in GeoBasics was simplified by making better use of local variables.