Avoid unneccessary ant runs for GWT
If the *.gwt.xml file is newer than the resulting *.nocache.js, then Make
will run the ant file, which will – if there were no other changes –
complete without modifying the already existing file.  As a result, the make
rule will trigger again in the next run.  Now we touch the file, thus
resetting its timestamp to the current time.

We also change the ANT_DEP to an order-only dependency: if ant gets updated,
we don't have to update all the GWT modules since there is no change in the
resulting files to be expected.