Fix bug in geoOps._helper.IntersectLC
It is wrong to always return the first row and column, since if one point of
intersection has an (almost) zero first coordinate, then the result would be
the null vector.
It is better to choose a maximally non-zero row resp. column.  This change
achieves that by using a new utility function to find the index of a list
element with maximal value.