Implement PointOnCircle.parameterPath in an attempt to control path
With this change, a movement of the mouse through the center of the circle
will at least result in a bad state and cause the point to flutter between
right and wrong position.  If it stabilizes then it will do so in the right
position.  But why it flutters, what causes it to stabilize and how that
stabilization might become instantaneous are all open questions, so this
commit is work in progress.