Change PointOnCircle.parameterPath to two arcs
Now the central point is real, which ensures that we have a uniquely defined
direction to move around the center of the circle.  Without this, the
complex tracing could lead to situations where things suddenly became
ambiguous, despite the fact that in the real interpretation the situation
was absolutely clear.

Judging from first experiments, this appears to work pretty well now.