Implement geoMacros, turn CircleMFixedr into a macro
This introduces a new mechanism “geoMacros” which allows defining some
geometric operations in terms of other geometric operations.  The return of
a macro function is a list of elements which will replace the matched
element in the gslp.  Macro expansion is recursive, so macros can build on
other macros.

The CircleMFixedr operation is now implemented as a macro, and will be
turned into a simple CircleMr which is pinned.  This also fixes a bug in the
previous CircleMFixedr implementation where the point p was relative to the
origin, not relative to the midpoint of the circle.