Drop code where computeParametersOnInput can detect third tracing step
Two tracing steps are enough to at least detect problems, at least for
PointOnCircle and judging from some manual experimentation.  Let's avoid
extra complexity, particularly since throwing an exception conflicts with
returning a value, and in the long run it would be nice if we could return
the new parameter instead of assigning it.