Reimplement PointOnLine
This changes the way how the point moves as the defining line changes
position.  Instead of orthogonal projection onto the line (which causes
points to become complex since the moment we decided to always do complex
detours), we now consider the circle around the intersection between old and
new line position, and a circle around that through the old point position.