Implement makeStatePermanent
As it was, a movement might store some state in stateOut which becomes
stateLastGood.  Then some other movement might swap states a number of
times, and during that time only copies and adjusts the state for the
elements depending on the mover.  So if the final state object is not the
same as the initial one for that move, then the final result of the
preceding move will be lost.  To avoid this, we copy the resulting state to
all buffers once a move is complete.