Introduce a “ports” attribute for createCindy
This new attribute can contain a list of ports, although at the moment we
only support a single such port.  Each port can be described by a number of
parameters, e.g. its id and dimensions.

The main benefit is that Cinderella can export all relevant data to a single
JSON object, and createCindy will take care of adjusting the dimensions of
the canvas element (which is hard-coded into the export template file).
That should help getting rid of a jQuery dependency there.