Trace position changes caused by scripts
Now changing the position of a free element via the script causes a tracing
very similar to that caused by mouse movement.

If the scripted movement is inside the move or draw event, then these
changes will be considered together with the mouse move: only if all of them
are good, then the resulting state is marked as good.  For assignments
outside the mouse movement code path, the state before is marked as good and
the resulting state is marked as good as well, effectively breaking tracing.