Clarify IntersectConicConic and solveCubic
The previous solveCubic implementation was confusing, so it's now demoted to
a helper.  Now a call to solveCubic will actually return the solutions, not
just something from which the solutions can be computed easily.  This moves
code from GeoOps to CSNumber, where it can be reused.

Computing the determinants by rows not columns makes no difference for the
result, since the matrices are assumed to be symmetric.  It does however
greatly simplify the code.  Numbering the coefficients of the cubic equation
makes the associated powers more readily apparent.

Normalizing for the maximal entry, instead of the z coordinate, allows us to
deal with points at infinity in a consistent way.