Simplify IntersectConicConic
Instead of four distinct cases, we now only handle two.
If both conics are degenerate, we use these directly,
otherwise we produce two distinct degenerate conics.

This has the benefit that there is only a single (and pretty rare)
situation where the threshold myeps is relevant.
Therefore, tuning that variable becomes less of an issue.