Move regression test to new directory
As an example, this file is pretty broing.  But as a regression test,
ensuring the bug demonstrated in f66987f6ca085c1f31aa2779be270f1204cc33e7
won't reappear, it might still make sense to keep this around.  A short
description helps evaluating the result, checking it for correctness.