Import resp. update KaTeX
This is from KaTeX commit 90ba2583 which provides render-to-canvas support.
It hasn't been merged into master yet, and if it gets merged, it will likely
be squashed.  After running "make", copy the build/katex directory from the
KaTeX source tree into the lib directory of CindyJS.

Since KaTeX is about to become a more integrated part of the platform,
I'm moving it out of the examples directory.