Improved KaTeX functionality
This is based off a merge of many unreviewed features which are currently
being developed and reviewed.  This makes a lot of features available, at
the risk of some instabilities.  We now load all the fonts up front, which
can take considerable time.  More work is needed to fix that aspect.