Update node to 0.12.6
This is the latest version of node.js at the moment. It might fix some bugs.
But more importantly, users will start having 0.12.x around more often,
so we should make sure that everything works with that version.
After all, asking them to upgrade is easy, for a downgrade less so.