Remove PopupConsoleHandler
As it stands, this handler is too annoying since it will unconditionally
open a Popup window, which in turn will get blocked by most modern web
browsers but cause an ugly warning message instead.  If that code is ever
required again, it could be revived from the git history, but since there is
currently no API to choose this handler, the code is effectively dead and
shouldn't be left to rot.