Deploy CindyJS to a GitHub repository using Travis CI
A new repository called CindyJS/deploy was created for this purpose.
Using a deploy key (travis-ssh), Travis CI has access to that repository.
The deploy key was encrypted using the travis command line tool.
See http://docs.travis-ci.com/user/encrypting-files/ for details.

The deployment itself works by exchanging the pre-existing content
of the deploy working tree with the output found in build/js.
This is done by simply moving those things which are to be maintained.
For details on the configuration of deployment and script support see
http://docs.travis-ci.com/user/deployment/
http://docs.travis-ci.com/user/deployment/script/