Improve two-ray case of drawarc
We don't need to extract point coordinates yet again, since we already know
the coordinates of a and c.  We can simply add to these to construct new
points, and by using suitable object literals we can avoid the previously
confusing name of that second point.

What we *do* need to do is consider the case where the user coordinate
system is zoomed out so far that 1000 isn't outside the canvas any more.
We have to take the scale of the current coordinate system into account.