Include each token in source map
Current versions of Chrome don't cope well with the kind of source map we
used to generate, with one entry per line.  Apparently it requires an exact
match for every token, or something like that.

The npm package `source-map-dummy` can be used to generate such a source
map.  And while we are increasing the number of external tools, we might
also add `convert-source-map` to do the actual concatenation, after some
massaging of the input to ensure that some sane expectations are being met.

Applying a stack of source maps becomes easier as well if we have an exact
match for every token, making some code there obsolete.