Make quickhull3d use the plugin API
So far we had one function in Operators.js turn CindyJS types into plain
arrays and back, and another in Java code to turn that into the
corresponding Java data structures.  Now we do the whole conversion in Java,
using CjsValue as an overlay type to represent CindyJS values.

This allows us to register the convexhull3d function as an operator provided
by a plugin, using the normal plugin API.  That way, the function will be
available with full functionality if the plugin has been loaded, or
completely unavailable if the plugin is not available.  No more calls to an
undefined function, as could happen before.