Auto-detect needed plugins based on usage of functions
When parsing scripts we keep track of used functions, with arity.  Later on
this information can be used to defer initialization (running the init
script in particular) until additional modules needed to supply these
functions have been loaded.

This is currently used for the convexhull3d function, provided by
GWT-compiled code and therefore loaded separately in CindyJS although it is
a built-in function in Cinderella.

Note that only the common set of scripts gets compiled during initialization
and treated this way.  Code encountered during cdy.evokeCS or similar will
run after initialization is already done, and won't trigger loading the
plugin.  If running such code is to be supported, the plugin should be
loaded explicitely.