Shift default origin by half a pixel
By shifting the default position of the origin, integer user coordinates
have a higher chance of ending up at the center of a pixel, as opposed to
the corner of a pixel.

This will affect constructions which use the default coordinate system, as
well as those which apply relative transformations to that default, like
`scale` and `translate`.  Absolute transformations like `scaleAndOrigin` and
`visibleRect` are not affected.