Fix measureText for FakeCanvas in unit tests
The `measureText` method of a `CanvasRenderingContext2D` returns a
`TextMetrics` object, not just the width.  The only thing we are using from
that object at the moment is the width, so it is enough to return that.

This explains and fixes a number of test cases where the `x` coordinate of
drawn text was `null` instead of some number.