Format angles as degrees by default
Now we're making use of the `usage: "Angle"` setting which might have been
applied to some numbers, and format the resulting numbers using degrees
instead of radians by default.  There is an instance argument called
`angleUnit` which might be used to override that default.  Various units are
already defined, using the turn as the unit of reference.  For future
extensions, it's already possible to enter a special formatting function
into that list.  Such a function could conceivably be used to e.g. format a
number as degrees, minutes and seconds.  That isn't supported yet, though.