Only update content when mouse actually moved
So far we run `updateCindy` continuously as long as the mouse button is held
down.  This can e.g. lead to traces decaying faster than they would in
Cinderella.  With these modifications, we only update the content if the
mouse actually moved, leading to improved compatibility and more efficient
use of resources.  The timer is still ticking all the time while the button
is down, which might be addressed at a later point.