Treat strings as literals when calling the replace operator
We have to call String.prototype.replace with a RegExp as its first argument
because we need the `g` (global) flag.  Otherwise we'd only be replacing the
first occurrence.  But we don't want to interpret the string as a regular
expression, but treat it literally.  So we have to backslash-escape any
characters which might have special meaning in a regular expression.  Which
means any non-alphanumeric character, just to be on the safe side.

We also have to make sure that the replacement string is interpreted
literally, without the special meaning of sequences starting in `$`.  We
achieve this by doubling any `$` encountered in that string.