Distinguish geometric elements by kind not type
The `kind` is the “output type” of a geometric operation: point, line,
conic, … as indicated by the comment at the beginning of GeoOps.  The `type`
is the operation itself, i.e. how that output is computed from its input.
Here only the kind is neccessary, since only that decides what additional
operations may be applied to it.  Since there is no dedicated kind for
circles, we can only check for conics.  The names should reflect that.