Avoid code duplication between csinit and addElement
Now all element creation is routed through addElement.  This means that we
don't know the total state size up front, but have to expand the state
incrementally as objects are added.  To avoid excessive reallocations, we're
reserving twice the newly needed size, so that we can extend the arrays
fairly cheaply until that capacity is exhausted.