Make drawing of traces more consistent with Cinderella
Now we implement the cubic alpha dimming and the dimension change in
the same way `CubicAlphaAgingRenderer` does things in Cinderella.

The property in the geometry list was renamed from `trace` to `drawtrace`.
That's what Cinderella exports already, and it's more descriptive and less
likely to be confused with the topic of complex tracing around
singularities.

The storage for past trace positions was changed from a list with endpoint
modifications to a list of fixed length with cyclical index.  Tests at
http://jsperf.com/shift-register/2 suggests that this should be orders of
magnitude better than what we had, and it also avoids trouble due to a
varying list length.