Distinguish text files from binary files
Explicitely making such a distinction will help possible contributors on
Windows to work with the code.  The new check in the Makefile ensures that
our list is exhaustive and stays that way.

The clipper README.txt had to be normalized in order to conform to the new
settings, since text files always use NL by itself in the repository.