Make sure to use the right version of npm for the top level install
We have some machinery to use a more recent version of node if the one
provided by the system is not recent enough.  But here was one place where
we failed to actually use that version.  One observation was on a Fedora
system with node 0.10.36 installed, where the system npm would install
release candidates of jshint, even though later versions of npm only install
prerelease versions if explicitely requested.