Allow different versions of node to coexist
Now we don't rename the extracted directory to download/node but leave the
full version and architecture specification in that directory name.  As a
consequence, multiple versions of node can now be extracted concurrently,
and it is enough to pass `NODE=false NODE_VERSION=4.2.0` or something like
that to make.  The `NODE=false` prevents the use of the system-installed
version of node.