Remove duplicates in List.js issue #88
This change removes the second occurrence of List.triples which is an exact duplicate. It has been been there since its introduction in @a1ef53ba91e39125def81c14aa2374594b152a00. 

Opportunistically it also removes the first occurrence of List._helper.compare which expects to be passed a number, whereas the second occurrence handles a List which is exactly what the only call to List._helper.compare in General.js passes to it:
```javascript
General.compare = function(a, b) {
...
    if (a.ctype === 'list') {
        return List._helper.compare(a, b);
    }
```