Implements CSNumber._helper.input
CSNumber._helper.input(a) can be used to convert a plain real number or an object of the form { r: number, i: number } to a CSNumber. Intended for use in allowing pos and other values to be input as complex numbers as well as real numbers. For example, pos: [ -3.0, -4.0, { r: -7.654042494670958E-17, i: 1.25 } ]