Honor the conceptual immutability of CS list values
To make copy-on-write work the way it does, we need to make sure that no
code modifies a list without copying it first.

Currently initializePoint will always return a new list, so modifying that
would be all right, technically speaking.  But conceptually this adds a
non-trivial constraint for future development of that helper, so it's best
to avoid that.