Fix Render2D.drawline to correctly clip
The test file demonstrates the problems. The clipping is not correct as currently implemented by Render2D.drawline. Depending on the angle of a line and where it is drawn it may be drawn outside the clipping rectangle and not within even though it clearly should be. The clipping rectangle is not being set from the port's visibleRect value. That has not been addressed here.