Clean up Pole and Polar, breaking compatibility
With PolarPoint and PolarLine, there was a horrible mess about which is
which, mostly because Cinderella uses them in an intuitive way: PolarPoint
is a line, PolarLine a point.  To avoid any such confusions in the future,
this commit introduces a new pair of nouns, namely Pole (which is a point
defined by a line and a conic) and Polar (which is a line defined by a point
and a conic).  The order was unifies so that the conic always comes last.

Since the state of affairs so far was such a mess, the commit deliberately
does not try to preserve backwards compatibility.