Run tasks in sequence unless requested otherwise
This introduces a new setting `parallel` to control the parallel or
sequential execution of tasks and jobs.  By default, tasks are now executed
sequentially, but jobs marked for parallel execution are executed in
parallel.  This may make it easier for developers to spot the error message
in case of a problem.  The `prepublish` script executed by `npm install`
still does parallel builds, though, sacrificing clarity of output for speed.