Design build description for repeated execution
Instead of relying on the exact order in which the different tasks are
executed, we now are back to requiring all modules at the top of the file
(with index.js as an unavoidable exception).  The build description now is a
function which takes the settings registry and the task factory as
arguments.  That way, multiple build runs can be executed with multiple
settings.

This commit contains a lot of whitespace changes, so you may want to use
`git show -w` (or `git diff -w`) to ignore changes in whitespace.  On GitHub
you can add `?w=1` to the URL in the address bar for the same effect.