Create hopefully persistent anchor names
While marked tries to provide anchor names for all its headings, these names
tend to focus on the word parts of the heading, ignoring symbolic content.
Changing the descriptive part of a heading line will change the anchor.  So
we override that, and try to come up with a scheme which will hopefully make
our anchors remain unaffected by many common forms of edits.

In particular, a normal CindyScript command has an anchor composed from the
name of the command and its arity.  An infix operator has its command
derived from the codepoint of its symbol(s).  Some other constructs like the
one for absolute value use `_` as a placeholder for arguments and unicode
codepoints for the non-argument symbols.  Only the first of these is really
easily readable by humans, but all of these should remain pretty unaffected
by changes to the non-essential parts of the headline.