Add .md extension to links to wiki pages
This will be useful for people who view the Markdown pages as rendered by
the GitHub source file browser, since it will allow them to navigate to the
desired page there.  It means that we will have to translate these
extensions when generating HTML pages, but that sounds pretty simple.