Replaces geoOps.ConicInSquare with PT version
This code implements geoOps.ConicInSquare using a projective transformation of the four corners of a square whose four sides are tangent to a unit circle into the four points given. The transformation is applied to the matrix of that unit circle resulting in the desired conic which is then tangent to the lines through those given points in the same manner as the square and the circle.