Change default text size from 20 to 12
This is to achieve compatibility with Cinderella.

The examples were adjusted to maintain their previous font sizes.  One
notable exception is examples/28c_Antiprism.html which benefits from the
change since its labels are now of the same size.

The textsize call has to happen in the move or draw script, since setting
the font size in the init script will not have any effect after the first
update.  This is due to the greset call in updateCindy, but it's the same
behavior as in Cinderella.