Increase font size in template
While compatibility with Cinderella mandates that the built-in default font
sizes should be 12px, that's pretty small for today's screens, or for beamer
use in education.  The size of 20px (previously used for drawtext calls)
makes more sense in such setups, so we might want to use that by default for
new examples.