Update geoOps.TrAffine.updatePosition
Found a way to get a list of z-values with having to resort to using List.column by getting the last row from the transpose of the point sets. Switched t(adj(ps1)) to adj(t(ps1)) to avoid unnecessary transpose to do so. Also collected duplicate mult calls into a vector to allow their reuse. So calls to multiMult are no longer needed and the scl function is now in-lined due to its simplicity.