Concise error message in case of a missing command
If a child process fails to spawn due to some missing command, the command
in question may not be part of the error message for some versions of node.
And even if it is mentioned, it may get lost in large stack traces.

It's better to detect this problem (which can be expected to occur quite
often if people lack some required dependency) and issue a shorter error
message without stack trace.