Prefix task output with task name
This makes the output easier to read, particularly if parallel=true has been
specified and the messages of different tasks get mixed.  The
WholeLineStream module ensures that we don't switch between tasks in the
middle of a line, and the attribution prefix allows us to tell what task a
given line belongs to.