Cindy3D: Support textured meshes
The support is very experimental at the moment. It will break if there are
combinations of meshes with and without texture, or using different
textures.  That's because the texture source property is global to the
instance, not set per mesh.  PrimitiveRenderer needs to be refactored, to
allow multiple sets of triangles associated with different textures, while
at the same time sharing the shader code.