Refer to whole-line-stream as a separate module
I found that I've been using this class in several places, often adding
support to one feature or another.  It became dufficult to track which
version had the latest code for which feature.  So now the whole line stream
is an npm module in its own right, and can be improved there and used here.