Auto-detect binary called nodejs
Debian and Ubuntu chose that name, in order to avoid a naming conflict.
Simply invoking a binary of that name won't work, though, since some
packages (like expect@1.13.4 at the moment) assume that they can invoke node
using its canonical name.  So we provide our own wrapper script to invoke
the node binary, and place that on the PATH to provide a command named node.

References:
https://lists.debian.org/debian-devel-announce/2012/07/msg00002.html
https://github.com/mjackson/expect/issues/64