Build debug version by default
The hard-to-remember `js_compiler` flag is replaced by a more general
`build=debug` vs. `build=release` setting.  The default is `debug`, so that
a simple `make` or `node make` will be optimized for developers who type
that command manually.  The prepublish script, on the other hand, uses the
release build mode, so users who are used to run `npm install` for any
package they depend on will get an optimized build.  This also holds for the
builds on Travis CI, while the tests (which also prepare for deployment)
require an explicit `build=release` flag which has been added.