Add example of AngularBisector Cinderella export
One occurrence of the primitive uses the current form with the third
parameter, while the other two use the shorter form with just the two lines.
So this file in its current form checks that both versions work as expected.
Based on AngularBisector (#172) @gagern

Modified to run lines a and b through separated points, create the intersection point of the two lines and  show the coordinates of the intersection point using CindyScript. The axes, grid and point snapping is also turned on.