Catch use of < and > to delimit placeholders in reference manual
The pattern is only detected if < and > delimit a word inside a code block
inside a headline.  Or something like this, since actually getting these
conditions perfectly right would require a markdown parser.

Also add some explanatory comments to the existing checks, to make them
easier to understand and maintain.