Change aspect ratio of arrowheads
So far, the arrowheads were equilateral triangles.  Now they are only
isosceles, with a ratio of 1:1 between length (i.e. extent along the line
segment) and breadth (i.e. length of the edge orthogonal to the baseline).
For comparison, since the length of the arrowhead is an altitude of the
triangle, that ratio used to be √3:2 in the equilateral heads.