Add arrow shapes empty and jet
This adds two new arrow shapes, `empty` (a.k.a. `hollow`) and `jet` (a.k.a.
delta).  It also introduces the name `line` for what used to be called
`default` up till now.  The latter name in each of these cases was in use by
Cinderella exports for a while, but is to be considered deprecated.

The mechanics is as follows: to draw an arrowhead of size 1 pointing in the
positive x direction and with the tip at (0, 0), we usually draw the curve

    (-ratio, -0.5) — (0, 0) — (-ratio, 0.5) — (-1, 0) — close

where ratio is 1 for most heads, but 1.5 for `jet`.  For `line` we only take
the first three vertices and don't close the path.  All the arrows are
defined by a set of flags which control the various aspects.