Implement the version function
At the moment, the version function will return a list

    ["CindyJS", 0, 0, 0, -1, "g‹commit-id›"]

which can be used to distinguish CindyJS from Cinderella, and to identify
the commit in question.  The commit id will be terminated by an exclamation
mark if the working directory is different from the named commit.

Once we start tagging releases, using annotated tags of the form

    v‹major›.‹minor›.‹patch›

then the second through fourth component will be the version components,
and the fifth component the number of commits since that version
(and therefore always 0 for official releases).
The machinery for this is already in place in the build system.