Introduce EuclideanMid as an alias to Mid
Cinderella has two distinct algorithms here, and in particular makes use of
EuclideanMid as part of measurement definitions.  So whenever a length is
measured, its midpoint gets constructed as an invisible point.  If a regular
midpoint gets constructed later on, the proover demonstrates the two to be
equal and preserves the older one.  This way, points constructed as
EuclideanMid may end up as visible points in addition to text anchors.