Fix Meet.visiblecheck
On the one hand, we now have segments which don't have type Segment but type
TransformS or some such.  So we have to distinguish by kind not type.

On the other hand, there was a copy and paste error in one of the indices,
so the wrong argument got referenced causing incorrect verdicts.