Change string representation of nada to `___`
This matches Cinderella, where Nada was always represented as `___`.
Presumably the distinction between `___` and `_?_` was to help investigate
problems with evaluation, but the exact association as to which is which
doesn't matter that much.  Apparently the deviation from Cinderella was
accidential, so reversing that will help developers switch between tools.