Fix mistake in HTML rendering of reference manual
The sequence `$&` in the input text, which might e.g. occur inside
`$&quot;` in the `TeX_Rendering` document, would be interpreted as a
replacement instruction by `String.replace`.  So we have to double all
`$` there, using yet another `String.replace` which of course needs all
`$` in its second argument doubled.