Make sure all parsed operators have SOME implementation
This orders the infixmap assignments equal to the operators, which makes it
easier to compare these two.  Unimplemented operators now will report an
error message on first invocation, and return nada.  This allows code using
such operations to parse successfully, so that a runtime case distinction
can be used to avoid actually relying on such computations.

The comments describing the operations were rephrased in English.