Let CindyJS accept the same set of variable names as Cinderella
Cinderella has a few special cases, and after that accepts everything which
has letter category (category L) in the Unicode 6.2.0 standard, is an ASCII
digit or an apostrophe.  Since Cinderella has no code to handle surrogate
pairs at this point, it is restricted to BMP letters.  So we encode that
into a regular expression.

There is one notable difference: Cinderella includes U+03A2 (decimal 930)
which lies within the Greek upper-case alphabet but is actually unassigned.