Reimplement Parser
The new parser is more modular and faster.
On my system the running time reported by tools/exampleScripts.js
dropped from 2.5s to 0.5s, so this is a factor 5 speed improvement.
The time complexity should be roughly linear, where it used to be roughly
quadratic before.

Note that the new parser does NOT implement exactly the same language as the
old one.  The changes are currently being discussed.