Approximate conic by quadratic Bézier curves
While cubic Bézier curves might be able to provide a better approximation
for a given number of control points, finding appropriate control points
involves solving a frouth-degree polynomial equation, and it is far from
clear which of the multiple results should be chosen.  Quadratic curves are
a lot easier, since all we need are tangent directions, and the solution is
unique.  So in the same time we can compute a lot more control points, more
than making up for the slightly inferior approximation quality.