Rename fuzzy operators
This changes the descriptions used for fuzzy comparison operators, in
documentation and source comments, to better match the order of symbols.

I'm doing this change against my own better judgement, as I personally
believe the new descriptions to be highly confusing.  But since @kranich and
@kortenkamp agree in their preference for this order, I'll relent and use
that.  See https://github.com/CindyJS/CindyJS/pull/248#discussion_r53288194
for details from the discussion.