Clarify namespace implementation
In the old parser, the variable object inside the parse tree was the same as
the object inside the namespace.  But almost no code actually made use of
this feature.  So the new parser doesn't have a stack in its variables,
while the new namespace has just the stacks, no ctype or other metadata.
This should help avoid confusion about which is which.